Position-based Draggable Images
Images now have to be in collision with the mouse to be dragged.
Mouse mode has to be set to absolute to do this- the bird will no longer follow the mouse properly in top-down mode, but as we shouldn't need a feature like this in the game, it shouldn't be a problem.