Event System is Functional
Complete rewrite from last push - now does not use different derived classes for the different event types, but instead uses an event class with a different payload depending on the type (uses a union and as such is not type safe). 

Future developments for the system include filtering events (only sending them to classes that request that kind of event).

Next I want to work on a json file that stores the keybindings, loading that in and using its data to check map input events to game events. Could add functionality to save any changes to this data (so that future projects could add an options page to change keybindings).