Basic Text On Remote
Remote now has Afterlife's text information on it.
Year and money strings are changed during tick to demonstrate the possibility. 

Envy and Avarice buttons display as selected when pressed, but looking for a more efficient way to give the rest of the buttons names than setting them one by one